remove ApolloMetadata
